Split views in different files
