+++ lemonpie/resumebuilder/views/__init__.py
@@ -1,0 +1,3 @@
+from .views_for_cv import *
+from .views_for_entry import *
+from .views_for_group import *
